prevent memory impl to access private fields + add missing pub/accessors
